+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Ignore `vendor` and `testdata` directories and directories with names
+  that begin with `.` or `_`
+
 ## [0.4.0] - 2021-12-19
 ### Added
 - Support method receivers with type parameters introduced in Go 1.18